Added FarPeekTextReader
The FarPeekTextReader wraps a regular TextReader with a small buffer to
allow peeking up to two characters ahead.

Peeking at two characters is useful for the parser because peeking at
one character might return '<', but the second character may or may not
be a ':' and reading the first could put the parser in an invalid state.

Someone needs to keep track of the extra character that needs to be
read, and a special object for the purpose seems like a good idea.