Temporarily bypassed reflection from TagFactory
TagFactory temporarily doesn't use reflection to construct tags now.
Reflection is fragile and complicated, and some significant changes are
going to need to be made to the Tag system, so a simpler-to-update
alternative will be used until the system re-solidifies.

Construction of user-defined tags now requires recompiling until the new
reflection method is built.